Move lifetime tracker into a separate module
